grid/prefix-sum refactor, more README
