+++ src/main/scala/org/hammerlab/magic/rdd/cmp/Unordered.scala
@@ -11,7 +11,8 @@ import math.{max, min}
  * Given an outer join of two [[RDD]]s with the presence or absence of values for a key replaced with a [[Boolean]] for
  * each [[RDD]], expose statistics about how many elements exist in either or both [[RDD]]s.
  */
-class Unordered[T: ClassTag] private(joined: RDD[(T, (Long, Long))]) {
+class Unordered[T: ClassTag] private(joined: RDD[(T, (Long, Long))])
+  extends Serializable {
   lazy val stats =
     joined
       .map {
@@ -44,7 +45,7 @@ class Unordered[T: ClassTag] private(joined: RDD[(T, (Long, Long))]) {
 }
 
 
-object Unordered {
+object Unordered extends Serializable {
   def apply[T: ClassTag](rdd1: RDD[T], rdd2: RDD[T]): Unordered[T] = {
     val keyedRDD1 = rdd1.map(_ → (1L, 0L))
     val keyedRDD2 = rdd2.map(_ → (0L, 1L))
