+++ src/main/scala/org/hammerlab/magic/rdd/grid/Message.scala
@@ -1,5 +1,8 @@
 package org.hammerlab.magic.rdd.grid
 
+/**
+ * Messages sent between partitions representing partial-sums
+ */
 sealed trait Message[T]
 case class BottomLeftElem[T](t: T) extends Message[T]
 case class LeftCol[T](m: Map[Int, T]) extends Message[T]
