+++ src/test/scala/org/hammerlab/magic/rdd/scan/ScanRDDTest.scala
@@ -6,6 +6,8 @@ import org.hammerlab.spark.test.suite.SparkSuite
 
 import scala.reflect.ClassTag
 
+import magic_rdds.ScanRDD
+
 trait ScanRDDTest
   extends SparkSuite {
